Update Request.php
Request->filter_xss returned an empty string if NULL was passed. With this fix it returns NULL.

This impacted in Request->post(), Request->get() and Request->param(), when a param did not exists, these functions were expected to return NULL by default, but when internally called filter_xss function and passing NULL, filter_xss() converted the NULL to "" when calling strip_tags().

With this change the expected result of Request->post(), Request->get() and Request->param() is correct.